+++ src/main/java/org/example/hugmeexp/domain/mission/service/MissionServiceImpl.java
@@ -46,19 +46,7 @@ import org.springframework.web.multipart.MultipartFile;
 public class MissionServiceImpl implements MissionService {
     private final MissionRepository missionRepository;
     private final MissionGroupRepository missionGroupRepository;
-    private final UserMissionRepository userMissionRepository;
-    private final UserMissionGroupRepository userMissionGroupRepository;
-    private final UserRepository userRepository;
     private final MissionMapper missionMapper;
-    private final UserMissionMapper userMissionMapper;
-    private final UserMissionSubmissionRepository userMissionSubmissionRepository;
-    private final UserMissionSubmissionMapper userMissionSubmissionMapper;
-    private final UserService userService;
-    private final MissionRewardExpLogRepository missionRewardExpLogRepository;
-    private final MissionRewardPointLogRepository missionRewardPointLogRepository;
-    private final UserMissionStateLogRepository userMissionStateLogRepository;
-    private final UserMissionStateLogMapper userMissionStateLogMapper;
-
     @Override
     @Transactional
     public MissionResponse createMission(MissionRequest missionRequest) {
@@ -145,208 +133,5 @@ public class MissionServiceImpl implements MissionService {
         return missions.stream()
                 .map(missionMapper::toMissionResponse)
                 .toList();
-    }
-
-    @Override
-    @Transactional
-    public UserMissionResponse challengeMission(String username, Long missionId) {
-        User user = userRepository.findByUsername(username)
-                .orElseThrow(UserNotFoundException::new);
-
-        Mission mission = missionRepository.findById(missionId)
-                .orElseThrow(MissionNotFoundException::new);
-
-        if (userMissionRepository.existsUserMissionByUserAndMission(user, mission)) {
-            throw new AlreadyExistsUserMissionException();
-        }
-
-        MissionGroup missionGroup = mission.getMissionGroup();
-
-        UserMissionGroup userMissionGroup = userMissionGroupRepository.findByUserAndMissionGroup(user, missionGroup)
-                .orElseThrow(UserMissionGroupNotFoundException::new);
-
-        UserMission userMission = UserMission.builder()
-                .user(user)
-                .mission(mission)
-                .userMissionGroup(userMissionGroup)
-                .progress(UserMissionState.NOT_STARTED)
-                .build();
-        userMissionStateLogRepository.save(UserMissionStateLog.builder()
-                .userMission(userMission)
-                .prevState(UserMissionState.NOT_STARTED)
-                .nextState(UserMissionState.NOT_STARTED)
-                .build());
-
-        return userMissionMapper.toUserMissionResponse(userMissionRepository.save(userMission));
-    }
-
-    @Override
-    @Transactional
-    public void changeUserMissionState(Long userMissionId, UserMissionState newProgress) {
-        UserMission userMission = userMissionRepository.findById(userMissionId)
-                .orElseThrow(UserMissionNotFoundException::new);
-        userMissionStateLogRepository.save(UserMissionStateLog.builder()
-                .userMission(userMission)
-                .prevState(userMission.getProgress())
-                .nextState(newProgress)
-                .build());
-        userMission.setProgress(newProgress);
-
-        userMissionRepository.save(userMission);
-    }
-
-    @Override
-    @Transactional
-    public void submitChallenge(Long userMissionId, SubmissionUploadRequest submissionUploadRequest, MultipartFile file) {
-        UserMission userMission = userMissionRepository.findById(userMissionId)
-                .orElseThrow(UserMissionNotFoundException::new);
-
-        if (userMissionSubmissionRepository.existsByUserMission(userMission)) {
-            throw new AlreadyExistsUserMissionSubmissionException();
-        }
-
-        String uploadDir = FileUploadUtils.getUploadPath(FileUploadType.MISSION_UPLOADS).toString();
-
-        Submission submission = userMissionSubmissionMapper.toEntity(submissionUploadRequest);
-
-        submission.setUserMission(userMission);
-
-        if (file == null || file.isEmpty()) {
-            throw new SubmissionFileUploadException("파일이 비어있거나 존재하지 않습니다.");
-        } else if (
-                file.getOriginalFilename() == null ||
-                        file.getOriginalFilename().isEmpty() ||
-                        !file.getOriginalFilename().contains(".")
-        ) {
-            throw new SubmissionFileUploadException("파일 확장자가 없습니다.");
-        }
-
-        String originalFilename = file.getOriginalFilename();
-
-        String safeFileName = FileUploadUtils.getSafeFileName(originalFilename);
-
-        // UUID를 사용하여 파일 이름을 안전하게 생성 및 중복 방지
-        String fileName = UUID.randomUUID().toString();
-        File destinationFile = new File(uploadDir, fileName);
-
-        submission.setFileName(fileName);
-        submission.setOriginalFileName(safeFileName); // 복원될 파일명
-
-        userMissionSubmissionRepository.save(submission);
-
-        userMission.setProgress(UserMissionState.IN_FEEDBACK);
-        // 파일 전송되기 전에 저장하고 파일 전송이 실패하면 롤백되므로(SubmissionFileUploadException)
-        // 고아 파일, 고아 레코드가 남지 않을 것임
-
-        try {
-            file.transferTo(destinationFile);
-        } catch (IOException e) {
-            throw new SubmissionFileUploadException("파일 저장 중 오류가 발생했습니다."); // RuntimeException을 던져서 IOException이 나더라도 롤백되게 함
-        }
-    }
-
-    @Override
-    public SubmissionResponse getSubmissionByMissionId(Long userMissionId) {
-        UserMission userMission = userMissionRepository.findById(userMissionId)
-                .orElseThrow(UserMissionNotFoundException::new);
-
-        Submission submission = userMissionSubmissionRepository.findByUserMission(userMission)
-                .orElseThrow(SubmissionNotFoundException::new);
-
-        return userMissionSubmissionMapper.toSubmissionResponse(submission);
-    }
-
-    @Override
-    @Transactional
-    public void updateSubmissionFeedback(Long userMissionId, SubmissionFeedbackRequest submissionFeedbackRequest) {
-        UserMission userMission = userMissionRepository.findById(userMissionId)
-                .orElseThrow(UserMissionNotFoundException::new);
-        Submission submission = userMissionSubmissionRepository.findByUserMission(userMission)
-                .orElseThrow(SubmissionNotFoundException::new);
-        submission.setFeedback(submissionFeedbackRequest.getFeedback());
-        userMission.setProgress(UserMissionState.FEEDBACK_COMPLETED);
-        userMissionRepository.save(userMission);
-        userMissionSubmissionRepository.save(submission);
-    }
-
-    @Override
-    @Transactional
-    public void receiveReward(Long userMissionId, String username) {
-        User user = userRepository.findByUsername(username)
-                .orElseThrow(UserNotFoundException::new);
-        UserMission userMission = userMissionRepository.findById(userMissionId)
-                .orElseThrow(UserMissionNotFoundException::new);
-        Mission mission = userMission.getMission();
-        if (userMission.getProgress() == UserMissionState.REWARD_RECEIVED) {
-            throw new AlreadyReceivedRewardException();
-        }
-        if (userMission.getProgress() != UserMissionState.FEEDBACK_COMPLETED) {
-            throw new InvalidUserMissionStateException();
-        }
-        userMissionStateLogRepository.save(UserMissionStateLog.builder()
-                .userMission(userMission)
-                .prevState(userMission.getProgress())
-                .nextState(UserMissionState.REWARD_RECEIVED)
-                .build());
-        userMission.setProgress(UserMissionState.REWARD_RECEIVED);
-
-        missionRewardExpLogRepository.save(
-                MissionRewardExpLog.builder()
-                        .userMission(userMission)
-                        .prevExp(user.getExp())
-                        .nextExp(user.getExp() + mission.getRewardExp())
-                        .note("미션 보상 경험치")
-                        .build()
-        );
-        userService.increaseExp(user, mission.getRewardExp());
-        missionRewardPointLogRepository.save(
-                MissionRewardPointLog.builder()
-                        .userMission(userMission)
-                        .prevPoint(user.getPoint())
-                        .nextPoint(user.getPoint() + mission.getRewardPoint())
-                        .note("미션 보상 구름조각")
-                        .build()
-        );
-        userService.increasePoint(user, mission.getRewardPoint());
-    }
-
-    @Override
-    public List<UserMissionStateLogResponse> getAllMissionStateLog(long userId, LocalDate startDate, LocalDate endDate) {
-        LocalDateTime startDateTime = startDate.atStartOfDay();
-        LocalDateTime endDateTime = endDate.plusDays(1).atStartOfDay();
-        return userMissionStateLogRepository.findByUserIdAndCreatedAtBetween(userId, startDateTime, endDateTime)
-                .stream().map(userMissionStateLogMapper::toUserMissionStateLogResponse).collect(Collectors.toList());
-    }
-
-    @Override
-    public UserMissionResponse getUserMission(Long missionId, String username) {
-        User user = userRepository.findByUsername(username)
-                .orElseThrow(UserNotFoundException::new);
-
-        Mission mission = missionRepository.findById(missionId)
-                .orElseThrow(MissionNotFoundException::new);
-
-        UserMission userMissions = userMissionRepository.findByUserAndMission(user, mission)
-                .orElseThrow(UserMissionNotFoundException::new);
-
-        return userMissionMapper.toUserMissionResponse(userMissions);
-    }
-
-    @Override
-    public List<UserMissionResponse> getAllUserMissionsByTeacher(String username) {
-        User teacher = userRepository.findByUsername(username)
-                .orElseThrow(UserNotFoundException::new);
-        List<UserMission> userMissions = userMissionRepository.findAllByMission_MissionGroup_Teacher(teacher);
-
-        return userMissions
-                .stream().map(userMissionMapper::toUserMissionResponse).collect(Collectors.toList());
-    }
-
-    @Override
-    public UserMissionResponse getUserMissionByChallengeId(Long challengeId) {
-        UserMission userMission = userMissionRepository.findById(challengeId)
-                .orElseThrow(UserMissionNotFoundException::new);
-
-        return userMissionMapper.toUserMissionResponse(userMission);
     }
 }
